requirements.txt
streamlit>=1.35
openai>=0.28.1
langchain>=0.0.208
pydantic>=2.0
python-dotenv>=1.0.0

